Make retrieval ranking depend on permission-safe embeddings
Slice A turns document retrieval from fixture-only matching into a provider-backed, JSON-vector ranking path while preserving offline determinism and leaving pgvector/cross-encoder work as explicit follow-up seams.

Constraint: Offline tests must remain deterministic without OPENAI_API_KEY, and retrieval must permission-filter before embedding ranking or future reranking.
Rejected: pgvector migration in this slice | storage/index migration is Slice B and needs separate approval.
Rejected: cross-encoder reranker now | owner wants it eventually, but it needs a provider/interface decision and should only see top-k authorized candidates.
Confidence: high
Scope-risk: moderate
Directive: Keep pgvector as first-stage candidate acceleration and cross-encoder reranking as second-stage over authorized top-k; never send unauthorized chunks to either stage.
Tested: uv run pytest tests/test_knowledge_ingestion.py tests/test_permission_aware_rag.py tests/test_conversations_api.py tests/test_agent_observability_evals.py -q (37 passed); uv run pytest -q (168 passed, 1 skipped); uv run ruff check . --no-cache; uv run ruff format --check .; git diff --check
Not-tested: Live OpenAI embedding API call; pgvector storage/search; cross-encoder reranker; browser demo smoke.

+++ my_agents/knowledge/routing.py
@@ -183,7 +183,7 @@ def answer_mode_for_route(
     return "general_knowledge"
 
 
-def is_relevant_retrieval_result(*, route: RetrievalRoute, source: str, score: int) -> bool:
+def is_relevant_retrieval_result(*, route: RetrievalRoute, source: str, score: float) -> bool:
     """Return whether one retrieved chunk should affect answer grounding."""
     if score <= 0:
         return False
